Fix segfault - Initialize diagnostic_updaters_ array
Previously the array was not initialized, so the contained diagnostic_updater::Updater* pointers were random - either nullptr or not. In the case the ptr was not nullptr, its initialization was skipped (didn't call new diagnostic_updater::Updater), causing segfault when updater->update() was called.

Also add a warning and fix print type